Cleaning up, preventing duplicate expense types, trying and failing with swipe to delete.

+++ index.ios.js
@@ -403,6 +403,11 @@ class Oikon2 extends Component {
   }
 
   onDeleteType(type) {
+    if (type && type.name === 'uncategorized') {
+      this.showErrorMessage('You can\'t delete "uncategorized".');
+      return false;
+    }
+
     try {
       DataDB.delete('type', type)
         .then(() => {
